+++ src/components/Welcome.jsx
@@ -1,0 +1,7 @@
+export default function Welcome({ username }) {
+  return (
+    <div className="welcome_container">
+      {`Welcome ${username}!`}
+    </div>
+  );
+};
